+++ lib/processor/branch.go
@@ -284,12 +284,12 @@ func newBranch(
 
 	var err error
 	if len(conf.RequestMap) > 0 {
-		if b.requestMap, err = bloblang.NewMapping("", conf.RequestMap); err != nil {
+		if b.requestMap, err = bloblang.NewMapping(conf.RequestMap); err != nil {
 			return nil, fmt.Errorf("failed to parse request mapping: %w", err)
 		}
 	}
 	if len(conf.ResultMap) > 0 {
-		if b.resultMap, err = bloblang.NewMapping("", conf.ResultMap); err != nil {
+		if b.resultMap, err = bloblang.NewMapping(conf.ResultMap); err != nil {
 			return nil, fmt.Errorf("failed to parse result mapping: %w", err)
 		}
 	}